+++ styled-block-builder.php
@@ -2,7 +2,7 @@
 /**
  * Plugin Name: Styled Block Builder
  * Plugin URI: https://github.com/IIP-Design/styled-block-builder
- * Description: Customizable animated blocks for WordPress sites
+ * Description: Customizable content blocks for WordPress sites
  * Version: v2.0.2
  * Author: Marek Rewers, U.S. Department of State, Global Public Affairs <designdevops@america.gov>
  * Text Domain: gpalab-blocks